adding basic 1-threshold function
I added the basic 1-threshold function, basing it off
of the built-in data-series filtering method. This could cost
run time as it is transforming the entire dataseries when
we may only want the first few entires. However, I thought
it simplified the code enought to be worth it on the
scale of data we are analyzing.

I also added a more explicit exception for the data
filing being absent as this is expected to happen
at some point.