window finding and no data bug fixes
In testing I found some small issue. First was
that the window inequality should
have been inclusive, and was skilling windows that
were exactly the right size. Second, In my final
logic to check if the last entry in the filtered
series was the end of a window,
I forgot to first confirm that any points had been
found within the thresholds at all. I added None
checks to handle this condition