refactoring window selecion to common routine
I refactored the window selection logic into a seperate
subroutine to simplify the overall code. There is one minor acceleration
that I left out here in for simplity vs speed. Technically, the first
3 routine only care about find the start of a valid window, but
are currently hoping off until the end of that window is found.
This code could be sped up by adding a greedy control bool
which would be false where we only care that we have found
a window, but that could be flipped to true for the final
routine that cares about finding fool windows.